markdown-survey/requirements.txt
streamlit==1.32.2
python-frontmatter==1.0.1
pyyaml==6.0.1
pandas==2.2.1
markdown==3.5.2 